# docs/api.yml
openapi: 3.0.0
info:
  title: 'WeDigital Garden - Core API'
  description: 'Documentation of Core API for We Digital Garden'
  version: 0.0.1
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/offices:
    get:
      tags:
        - Office
      summary: 'Display a collection of offices ressources.'
      operationId: 'App\Http\Controllers\Api\OfficeController::index'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
      responses:
        '200':
          description: 'A offices collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/office_collection_resource'
        '404':
          description: 'A offices collection not found'
  '/api/offices/{id}':
    get:
      tags:
        - Office
      summary: 'Display the specified office resource.'
      operationId: 'App\Http\Controllers\Api\OfficeController::getOne'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: office_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
      responses:
        '200':
          description: 'One office'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/office_resource'
        '404':
          description: 'A office not found'
...
  /api/social-media:
    get:
      tags:
        - 'Social Media'
      summary: 'Display a collection of social media ressources.'
      operationId: 'App\Http\Controllers\Api\SocialMediaController::index'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
      responses:
        '200':
          description: 'A social media collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/social_media_collection_resource'
        '404':
          description: 'A social media collection not found'
  '/api/social-media/{id}':
    get:
      tags:
        - 'Social Media'
      summary: 'Display the specified social media resource.'
      operationId: 'App\Http\Controllers\Api\SocialMediaController::getOne'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the social media to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
      responses:
        '200':
          description: 'One social media'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/social_media_resource'
        '404':
          description: 'A social media not found'
...
components:
  schemas:
    resource_id:
      description: 'The unique identifier of a resource'
      minimum: 1
      example: '1'
...
    office_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/office_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/offices
          type: object
      type: object
    office_resource:
      required:
        - type
        - id
        - attributes
        - links
      properties:
        type:
          description: 'Transform the resource into an array.'
          type: string
          example: offices
        id:
          $ref: '#/components/schemas/resource_id'
        attributes:
          description: 'Transform the resource into an array.'
          required:
            - address
            - photo
            - link_gmap
          properties:
            address:
              description: 'Transform the resource into an array.'
              required:
                - address
                - postal_code
                - city
                - country
              properties:
                address:
                  description: 'The address of office'
                  type: string
                  example: '103 rue de Grenelle'
                postal_code:
                  description: 'The postal code of office'
                  type: string
                  example: '75007'
                city:
                  description: 'The city of office'
                  type: string
                  example: Paris
                country:
                  description: 'The country of office'
                  type: string
                  example: France
              type: object
            photo:
              description: 'Transform the resource into an array.'
              required:
                - url
                - custom_properties
              properties:
                url:
                  description: 'The url of photo of office'
                  type: string
                  example: 'http://localhost:4646/storage/media/1/0b024b6d74c506b423d90f47d28438e3.jpg'
                custom_properties:
                  description: 'The custom properties of photo of office'
                  type: object
                  example: '{''description'': ''photo of Paris office''}'
              type: object
            link_gmap:
              description: 'The google map link of office'
              type: string
              example: 'https://www.google.fr/maps/place/103+rue+de+Grenelle,75007+Paris'
          type: object
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/offices/1
          type: object
      type: object
...
    social_media_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/social_media_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/social-media
          type: object
      type: object
    social_media_resource:
      required:
        - type
        - id
        - attributes
        - links
      properties:
        type:
          description: 'Transform the resource into an array.'
          type: string
          example: 'social media'
        id:
          $ref: '#/components/schemas/resource_id'
        attributes:
          description: 'Transform the resource into an array.'
          required:
            - name
            - link
          properties:
            name:
              description: 'The name of social media'
              type: string
              example: LinkedIn
            link:
              description: 'The link of social media'
              type: string
              example: 'https://www.linkedin.com'
          type: object
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/social-media/1
          type: object
      type: object
...
  securitySchemes:
    bearer:
      type: http
      name: bearerAuth
      bearerFormat: JWT
      scheme: bearer